requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
pydantic>=2.11,<3.0
pydantic[email]>=2.11,<3.0
email-validator>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
GDAL==3.6.2
gunicorn>=21.0,<22.0